Moving testcase to utils
